// webSyscsc/ClientApp/src/app/syscsc/models/insumo-interno.ts
export class InsumoInterno {
    insumoInternoId: number;
    productorId: number;
    nombre: string;
    materialesUsado: string;
    procedimiento: string;
    tiempoPreparacion: string;
    metodoPreparacion: string;
    dosis: string;
    cantidad: string;
    fechaAplicacion: string;
    lugarAplicacion: string;
}